added bluff body + added example toy aircraft

+++ tbx/+baff/+model/@Hinge/ToBaff.m
@@ -13,7 +13,7 @@ function ToBaff(obj,filepath,loc)
     h5write(filepath,sprintf('%s/Rotation',loc),[obj.Rotation],[1,1],[1,N]);
     h5write(filepath,sprintf('%s/K',loc),[obj.K],[1,1],[1,N]);
     h5write(filepath,sprintf('%s/C',loc),[obj.C],[1,1],[1,N]);
-    h5write(filepath,sprintf('%s/isLocked',loc),[obj.isLocked],[1,1],[1,N]);
+    h5write(filepath,sprintf('%s/isLocked',loc),double([obj.isLocked]),[1,1],[1,N]);
 
     h5writeatt(filepath,[loc,'/'],'Qty', length(obj));
 end
